Fall back to service.port when httpPort tag is missing, and ignore missing context-base

+++ ob1k-consul/src/main/java/com/outbrain/ob1k/consul/HealthInfoInstance.java
@@ -48,7 +48,7 @@ public class HealthInfoInstance {
     public String Service;
     public Set<String> Tags;
     public String Address;
-    public long Port;
+    public int Port;
     public long ModifyIndex;
 
     public Integer port(final String portType) {
